requirements.txt
streamlit
requests
beautifulsoup4
langchain-groq
python-dotenv
tavily-python